refactor(coding-agents): split claude/pi handlers into their own files
plugin.go kept both Claude's host/workspace handlers under generic
names (hostMode, workspaceMode) alongside pi's agent-prefixed pair,
which forced readers to infer "this is Claude" from context.

- Rename hostMode/workspaceMode to claudeHostMode/claudeWorkspaceMode
  and extract a handleClaude dispatcher that mirrors handlePi.
- Move Claude handlers into claude.go and pi handlers into pi.go.
- plugin.go now holds only the Plugin struct, PreContainerRun
  dispatcher, and getCredentialsMode.

No behavior change; tests unchanged.
